requirements.txt
fastapi[all]>=0.115.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.31.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
slowapi==0.1.9
httpx==0.25.2
aiosqlite==0.19.0
fastapi-mail>=1.6.0
black==23.11.0
ruff==0.1.6

